+++ tests/testdata/id_rsa
@@ -1,0 +1,27 @@
+-----BEGIN OPENSSH PRIVATE KEY-----
+b3BlbnNzaC1rZXktdjEAAAAABG5vbmUAAAAEbm9uZQAAAAAAAAABAAABFwAAAAdzc2gtcn
+NhAAAAAwEAAQAAAQEAwUx4uOwEnUngBtygJ97XVIQ4H7tE22El9wdHh5wy9j1l1m2CVJHg
+AAghYvid9Uwns00wWK3fNUMc1XpUtKF9SvA42jDp2WUZST2qjcdutXc/HM65p5e2tFT5+8
+PupPl6l6YuaITW33mqlB9AJsJxX6jb+cNYl6IbcoqbVA4QBcr8S2zitcWyPl3brxfU3k6W
+jKQfcZELmINEwVez1y+Phx3XOzJLs3k9RY61wuQaRWvmJXYdzG6wXSAYpEou0GQQsW6bXg
+TpW1araSX/csb27N+GdbKz1xju9Nk/9ydO0otyoImNSCDq7oZHydP+TUYHGa0yShI6dLIv
+jZXVb5gSzQAAA8CO8/tojvP7aAAAAAdzc2gtcnNhAAABAQDBTHi47ASdSeAG3KAn3tdUhD
+gfu0TbYSX3B0eHnDL2PWXWbYJUkeAACCFi+J31TCezTTBYrd81QxzVelS0oX1K8DjaMOnZ
+ZRlJPaqNx261dz8czrmnl7a0VPn7w+6k+XqXpi5ohNbfeaqUH0AmwnFfqNv5w1iXohtyip
+tUDhAFyvxLbOK1xbI+XduvF9TeTpaMpB9xkQuYg0TBV7PXL4+HHdc7MkuzeT1FjrXC5BpF
+a+Yldh3MbrBdIBikSi7QZBCxbpteBOlbVqtpJf9yxvbs34Z1srPXGO702T/3J07Si3KgiY
+1IIOruhkfJ0/5NRgcZrTJKEjp0si+NldVvmBLNAAAAAwEAAQAAAQEAjohZ+FmoqWag985V
+ljC+bQa178gB5RZEjVvjxb5K4cGyRyQm1z9dggh8veL/brJCD/eZ77n9/eNH7q4uqabuoY
+RzF4i71t6yf4wmKDdnjg2pDw/sXAUF5wFF9/ypCyMe35XdYJqbatvnCSPmBO6Z9OSpmAkQ
+WXdib9p+4MMh03JqYdQCG25aT+qkdtLD16UCyGkXDWaARP+aZd4+ncB5FQhzJHIwQ9mmrE
+zBuyRHLZ1KEYxlDy59A22BNrViB0hu+1uP9DICu+OjhyHXVKrkLSWuA7Jg/gysHEXsURfZ
+thft+40GaNomWdSNZzK2Ex8F+V2vGQ7ir08iWXBATR2JwQAAAIBr/jaUHDicy5JYCjM7UQ
+7T9avXi1Z3HkJ4LtC7yKAWAw8kp38+itibQU3DedE+VX9BLdo4GnIyUChT/m+bdHysiHfU
+aonlHvj4iTUo5l89uCx+/ftRKyDt1g+cqhq4eKS1Q0PldBFamUrZNlZz1lyK6U6cZF+Tyo
+DADzgSsCXwhgAAAIEA7DZM28KlxC6sJQuUq+qUj/3Ue8XHIi2WNZ4UVTJgHV3K0xrI6KUK
+94VCYBOEb93IqclIbX2cPVuRPTWpBb3htydjCpnQ08tCN4tVvFH+jyxUVSm2pb1L449UJx
+4WA+UIkR+CfkiQ8CqEPY7VM7Odav+Yi1w7KDO0bF6kTmSeD1EAAACBANF93qwzqUtoERZE
+HqMVO+FPJDtrbUg1Nzdb4Qp3kyas13UFfPp+qyikUI/aDM5pitDvAtsAnjRhQQX9qG4ReU
+YX5dcxNEvnnPy8/8MbWLFzQ9slvtLFry0zMZ3RyxIN8xSuJr4xAC3FDrBnFIaEoenMOGhR
+1yhNBy7iEvYQsYS9AAAACnRob3JASGFkZXM=
+-----END OPENSSH PRIVATE KEY-----
